Add Support for cURL Snippets with Line Breaks
To add support for cURL snippets with line breaks, a few
classes were added:

- `CurlPart` - represents a part of a cURL command
- `CurlLineGroup` - a group of cURL curl parts that make up a line
- `CurlLineBreakStrategy` - a strategy for how lines should
be broken up
- `CurlLineBreakStrategies` - factory methods for strategies

To support configurability, the `CurlRequestSnippet` has
been almost completely redone. Methods that originally
wrote out a part of a command, have been abstracted into
a into a class, for which instances can be mapped to
the according `CurlPart`. Users also have the ability
to create their own strategy, and simply plug it in
through an operation attribute.

Tests have been also added for all strategies and custom
user strategies.

[gh-260](https://github.com/spring-projects/spring-restdocs/issues/260)