Provide a default output directory for the snippets. Fixes gh-297.
The path <projectDirectory>/<gradle/mavenDirectory>/generated-snippets will be
respected by a) a new block macro and b) the spring-restdocs-core class ManualRestDocumentation.
This behaviour can be overridden by a system property called snippetOutputDirectory.

a) The new block macro module (spring-restdocs-asciidoctor-extensions) is a dependency
for asciidoctors maven or gradle plugin. It provides the class RestDocsSnippetBlockMacro
that includes and let asciidoctor process the content from the default (or custom)
folder if the macro restdocs:: is used.

b) The ManualRestDocumentation class and the classes using it as a delegate were
modified. You cannot provide the documentation path as constructor parameter anymore.
Instead it uses the same mechanisms as a).

All tests (also the tests of the samples) were changed to the new signature.

There is no change in the documentation. This will happen if this commit / pull request
got reviewed.