Upgrade Travis configuration to use a recent version of Oracle JDK 8
By default, Travis uses a rather old version of Oracle's JDK 8 in
its container-based infrastructure. It would appear that this is what
has been causing the build to fail with minimal diagnostics.

This commit implements a suggestion [1] to get a more recent version
of Java 8 by installing the oracle-java8-installer package via the APT
addon.

[1] https://github.com/travis-ci/travis-ci/issues/3259#issuecomment-130860338