Fix handling of empty querty string parameters
Previously empty parameters in a query string were handled
inconsistently such that they sometimes had a value of "" in a list
and sometimes were represented as an empty list. This inconsistency
lead to the parameter appearing twice in the cURL request snippet.

This commit removes the inconsistency by always using an empty string
to represent an empty query parameter value.

Fixes gh-647