Ensure that default value for snippets attribute is an absolute path
Previously, the snippets attribute was absolute when using Gradle and
relative to the docdir when using Maven. The relative path that was
used with Maven only worked as long as the working directory when
invoking Asciidoctor was the same as the docdir. This was the case
until 3.1.0 of the Asciidoctor Maven Plugin when the working
directory and docdir diverged at which point the snippets could no
longer be found.

This commit updates the snippets directory resolver to return an
absolute file when using Maven, just has it already does when using
Gradle. The resolution for Gradle has also been updated to explicity
make the File absolute rather than relying on the gradle-projectdir
or projectdir attributes having an absolute value.

Fixes gh-950