Introduce Toolchain plugin
Nullability support requires us to compile with Java 23 or later. We
still need to be able to test against earlier versions down to our
Java 17 baseline.

This commit introduces a Toolchain plugin that is apply by convention
to all Java project. When the toolchainVersion property is
configured, the plugin configures test tasks to use that Java version.